Day 21: Rust, with logic from Reddit
I quickly saw that this day's tasks can be seen as another instance of a graph and solved with
shortest-path / dijkstra.

I tried preparing  all shortest-paths, but realized afterward due the need of recursion,
this cant be pre-pared and for each button press alone a shortest-path for every layer
need to be found. This includes doing dijkstra on each and every level for each button press.
(Remark: A rust-solution for this approach is in [1])

I kinda was too much stuck before realizing this, thus tried to find a whole other solution
looking into Reddit. Some solutions had kind of shortest-path-rules, e.g. avoid zig-zag,
but I was unable to adopt my present code, without the recursive dijkstra, to work fully.

Therefore, I go here with another solution, also found on Reddit,  where the heavy lifting
is done as manual calculated shortest-path map by the author of that solution.